+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -234,6 +234,7 @@ object HybrisIcons {
     val CCV2_BUILD_STATUS_FAIL = ExpUiIcons.Run.TestFailed
     val CCV2_BUILD_STATUS_DELETED = ExpUiIcons.Run.KillProcess
     val CCV2_BUILD_CREATE = AllIcons.Actions.BuildAutoReloadChanges
+    val CCV2_BUILD_REDO = AllIcons.Actions.BuildAutoReloadChanges
 
     val CCV2_ENV_DEPLOYMENT_STATUS_SCHEDULED = ExpUiIcons.Run.Profile
     val CCV2_ENV_DEPLOYMENT_STATUS_DEPLOYING = ExpUiIcons.Run.TestCustom